Added better url sanitization Fixed bugs
