+++ .travis.yml
@@ -28,7 +28,7 @@ before_install:
   - docker network create web
 
 script:
-  - pytest discord/variable_store.py
+  - pytest discord/test.py
   - docker-compose build
   - docker-compose up -d
